mt5-service/requirements.txt
MetaTrader5==5.0.5200
python-dotenv==1.0.0
redis==5.0.1
pika==1.3.2
colorama==0.4.6